+++ src/ideabox/restapi/form/action.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from ideabox.policy.content.campaign import ICampaign
+from ideabox.policy.content.priority_action import IPriorityAction
 from ideabox.restapi import _
 from imio.restapi.form.action import RESTAction
 from zope.component import adapter
@@ -13,3 +14,11 @@ class PSTActionImportAction(RESTAction):
     application_id = u"PST"
     schema_name = None
     view_name = "pstaction-import-form"
+
+
+@adapter(IPriorityAction, IBrowserRequest)
+class PSTActionImportProgress(RESTAction):
+    title = _(u"Import PST Progresses")
+    application_id = u"PST"
+    schema_name = None
+    view_name = "pstaction-progress-import-form"